Improve version display
