+++ cli/util/text.go
@@ -3,6 +3,7 @@ package util
 import (
 	"fmt"
 	"os"
+	"runtime/debug"
 
 	"github.com/fatih/color"
 )
@@ -47,13 +48,16 @@ func Error(message string, exitCode int) {
 func BuildVersion(version, commit, date, builtBy string) string {
 	result := version
 	if commit != "" {
-		result = fmt.Sprintf("%s, commit: %s", result, commit)
+		result = fmt.Sprintf("%s\ncommit: %s", result, commit)
 	}
 	if date != "" {
-		result = fmt.Sprintf("%s, built at: %s", result, date)
+		result = fmt.Sprintf("%s\nbuilt at: %s", result, date)
 	}
 	if builtBy != "" {
-		result = fmt.Sprintf("%s, built by: %s", result, builtBy)
+		result = fmt.Sprintf("%s\nbuilt by: %s", result, builtBy)
+	}
+	if info, ok := debug.ReadBuildInfo(); ok && info.Main.Sum != "" {
+		result = fmt.Sprintf("%s\nmodule version: %s, checksum: %s", result, info.Main.Version, info.Main.Sum)
 	}
 	return result
 }
